requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
ping3==4.0.4
pythonping==1.1.4
APScheduler==3.10.4
python-dotenv==1.0.1
mysql-connector-python==8.4.0
PyMySQL==1.1.1
cryptography==42.0.8
gunicorn==22.0.0
